Dropped support for compiling FSTs.
This drops support for compiling FSTs from Ruby. SFST lacks an interface
for this so the existing code included the compiler and lexer code from
SFST upstream. The upstream code for compiling FSTs is in
(utf8-)scanner.ll and fst-compiler.yy, where it is intertwined with code
for the command-line utility fst-compiler(-utf8) using a lot of global
variables and other nastyness. This makes maintenance quite bothersome
and is the reason why this gem has been stuck at an obsoleted version of
SFST for some time. It seems to me that the sensible thing is to drop
support for this functionality until SFST exposes this functionality
properly.